feat: searchbar: move funtion/const outside 'Accueil'
